requirements.txt
skorch
torch
numpy
pandas
Pillow
streamlit